requirements.txt
# Web Scraping & API
requests==2.31.0
beautifulsoup4==4.12.3
tweepy==4.14.0
praw==7.8.1
googlenews==1.6.15
newsapi-python==0.2.7
selenium==4.16.0
webdriver-manager==4.0.1

# Data Processing
pandas==2.1.4
numpy==1.24.3
langdetect==1.0.9

# NLP & ML
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Environment & Utils
python-dotenv==1.0.0
python-dateutil==2.8.2

# Rate Limiting
ratelimit==2.2.1

# Visualization
plotly==5.18.0

# Web Framework
flask==3.0.0
